recording working wahoo!!
note, I switched the buzzer pin to an unused pin during dev work. This will need to be changed back to 8 for it to work with v01 hardware.

debounce on buttons working pretty good with initial testing. Went with a "wait for release" style. This may need to be tweaked in the future, but seems good as is for now.